+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Spark export: add Spark SructType exporter (#277)
 
 ### Changed
 
@@ -17,7 +18,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `datacontract serve` start a local web server to provide a REST-API for the commands
 - Provide server for sql export for the appropriate schema (#153)
 - Add struct and array management to Glue export (#271)
-- Spark export: add Spark SructType exporter (#277)
 
 ### Changed
 - Introduced optional dependencies/extras for significantly faster installation times. (#213)