+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Spark export: add Spark SructType exporter (#277)
+- add `--schema` option for the `catalog` and `export` command to provide the schema also locally
 
 ### Changed