+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- add `--schema` option for the `catalog` and `export` command to provide the schema also locally
+
+### Changed
+
+## [0.10.8] - 2024-06-19
+
+### Added
 - `datacontract serve` start a local web server to provide a REST-API for the commands
 - Provide server for sql export for the appropriate schema (#153)
 - Add struct and array management to Glue export (#271)